requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.7
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
jinja2==3.1.2
aiofiles==23.2.1
